Move event locker to writeValue method
