+++ src/core/component.ts
@@ -101,9 +101,7 @@ export abstract class DxComponent implements OnChanges, OnInit, DoCheck, AfterCo
         }
 
         if (this.instance) {
-            this.eventHelper.lockedEventName = name + 'Change';
             this.instance.option(name, value);
-            this.eventHelper.lockedEventName = '';
         } else {
             this._initialOptions[name] = value;
         }
